feat: allow multiple simultaneous clients (#10)
* fix: hard sync client state with server

This involves no prediction or optimistic updates
of any sort on the client.

* build: redirect cli arguments to client

* feat(mcp): add Try* versions of Send and Receive

* feat: accept multiple clients but with same state

* add todos

* feat(client): lerp between last two snapshots

* fix(server): run at 10 tps

* fix: drop old states using index

* perf(mcp): reduce lock critical section

* fix(mcp): ensure select opts of already sent chans

The default branch is kind of useless since open is only ever useful
if the selected branch is a receive operation (ctx and ln.die). This is
also why the concern of a panic is noted for the send branches.

* chore: remove debug logs

* docs(simulation): note why two separate for-loops are needed